patch: simplify resume upload code
